refactor pod gc
