+++ pkg/checker/podstartup/pods_test.go
@@ -2,6 +2,7 @@ package podstartup
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"testing"
 	"time"
@@ -11,7 +12,9 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/validation"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	k8sfake "k8s.io/client-go/kubernetes/fake"
+	k8stesting "k8s.io/client-go/testing"
 )
 
 func TestGenerateSyntheticPod(t *testing.T) {
@@ -203,5 +206,142 @@ func TestPodStartupChecker_getSyntheticPodIP(t *testing.T) {
 			podIP, err := checker.getSyntheticPodIP(context.Background(), podName)
 			tt.validateRes(g, podIP, err)
 		})
+	}
+}
+
+func TestPodStartupChecker_syntheticPodgarbageCollection(t *testing.T) {
+	checkerName := "chk"
+	syntheticPodNamespace := "checker-ns"
+	checkerTimeout := 5 * time.Second
+	syntheticPodLabelKey := "cluster-health-monitor/checker-name"
+
+	tests := []struct {
+		name        string
+		client      *k8sfake.Clientset
+		validateRes func(g *WithT, pods *corev1.PodList, err error)
+	}{
+		{
+			name: "only removes pods older than timeout",
+			client: k8sfake.NewClientset(
+				podWithLabels("chk-synthetic-old", syntheticPodNamespace, map[string]string{syntheticPodLabelKey: checkerName}, time.Now().Add(-2*time.Hour)),
+				podWithLabels("chk-synthetic-new", syntheticPodNamespace, map[string]string{syntheticPodLabelKey: checkerName}, time.Now()),
+			),
+			validateRes: func(g *WithT, pods *corev1.PodList, err error) {
+				g.Expect(err).NotTo(HaveOccurred())
+				g.Expect(pods.Items).To(HaveLen(1))
+				g.Expect(pods.Items[0].Name).To(Equal("chk-synthetic-new"))
+			},
+		},
+		{
+			name: "no pods to delete",
+			client: k8sfake.NewClientset(
+				podWithLabels("chk-synthetic-too-new", syntheticPodNamespace, map[string]string{syntheticPodLabelKey: checkerName}, time.Now()), // pod too new
+				podWithLabels("chk-synthetic-no-labels", syntheticPodNamespace, map[string]string{}, time.Now().Add(-2*time.Hour)),              // old pod wrong labels
+				podWithLabels("no-name-prefix", syntheticPodNamespace, map[string]string{}, time.Now().Add(-2*time.Hour)),                       // pod missing name prefix
+			),
+			validateRes: func(g *WithT, pods *corev1.PodList, err error) {
+				g.Expect(err).NotTo(HaveOccurred())
+				g.Expect(pods.Items).To(HaveLen(3))
+				actualNames := make([]string, len(pods.Items))
+				for i, pod := range pods.Items {
+					actualNames[i] = pod.Name
+				}
+				g.Expect(actualNames).To(ConsistOf([]string{"chk-synthetic-too-new", "chk-synthetic-no-labels", "no-name-prefix"}))
+			},
+		},
+		{
+			name: "only removes pod with checker labels",
+			client: k8sfake.NewClientset(
+				podWithLabels("chk-synthetic-pod", syntheticPodNamespace, map[string]string{syntheticPodLabelKey: checkerName}, time.Now().Add(-2*time.Hour)),
+				podWithLabels("chk-synthetic-no-label-pod", syntheticPodNamespace, map[string]string{}, time.Now().Add(-2*time.Hour)),
+			),
+			validateRes: func(g *WithT, pods *corev1.PodList, err error) {
+				g.Expect(err).NotTo(HaveOccurred())
+				g.Expect(pods.Items).To(HaveLen(1))
+				g.Expect(pods.Items[0].Name).To(Equal("chk-synthetic-no-label-pod"))
+			},
+		},
+		{
+			name: "error listing pods",
+			client: func() *k8sfake.Clientset {
+				client := k8sfake.NewClientset()
+				client.PrependReactor("list", "pods", func(action k8stesting.Action) (handled bool, ret runtime.Object, err error) {
+					// fail the List call in garbageCollect because it uses a label selector. This prevents breaking the test which also
+					// lists pods but does not use a selector.
+					listAction, ok := action.(k8stesting.ListAction)
+					if ok && listAction.GetListRestrictions().Labels.String() != "" {
+						return true, nil, errors.New("error bad things")
+					}
+					return false, nil, nil
+				})
+				return client
+			}(),
+			validateRes: func(g *WithT, pods *corev1.PodList, err error) {
+				g.Expect(err).To(HaveOccurred())
+				g.Expect(err.Error()).To(ContainSubstring("failed to list pods for garbage collection"))
+			},
+		},
+		{
+			name: "error deleting pod",
+			client: func() *k8sfake.Clientset {
+				client := k8sfake.NewClientset(
+					podWithLabels("chk-synthetic-pod-1", syntheticPodNamespace, map[string]string{syntheticPodLabelKey: checkerName}, time.Now().Add(-2*time.Hour)),
+					podWithLabels("chk-synthetic-pod-2", syntheticPodNamespace, map[string]string{syntheticPodLabelKey: checkerName}, time.Now().Add(-2*time.Hour)),
+				)
+				// only fail the Delete call for old-pod-1
+				client.PrependReactor("delete", "pods", func(action k8stesting.Action) (handled bool, ret runtime.Object, err error) {
+					deleteAction, ok := action.(k8stesting.DeleteAction)
+					if ok && deleteAction.GetName() == "chk-synthetic-pod-1" {
+						return true, nil, errors.New("error bad things")
+					}
+					return false, nil, nil
+				})
+				return client
+			}(),
+			validateRes: func(g *WithT, pods *corev1.PodList, err error) {
+				g.Expect(err).To(HaveOccurred())
+				g.Expect(err.Error()).To(ContainSubstring("failed to delete old synthetic pod"))
+				g.Expect(pods.Items).To(HaveLen(1)) // one pod should be deleted
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			g := NewWithT(t)
+
+			checker := &PodStartupChecker{
+				name: checkerName,
+				config: &config.PodStartupConfig{
+					SyntheticPodNamespace:      syntheticPodNamespace,
+					SyntheticPodLabelKey:       syntheticPodLabelKey,
+					SyntheticPodStartupTimeout: 3 * time.Second,
+					MaxSyntheticPods:           5,
+				},
+				timeout:      checkerTimeout,
+				k8sClientset: tt.client,
+			}
+
+			// Run garbage collect
+			err := checker.syntheticPodGarbageCollection(context.Background())
+
+			// Get pods for validation
+			pods, listErr := tt.client.CoreV1().Pods(syntheticPodNamespace).List(context.Background(), metav1.ListOptions{})
+			g.Expect(listErr).NotTo(HaveOccurred())
+
+			tt.validateRes(g, pods, err)
+		})
+	}
+}
+
+func podWithLabels(name string, namespace string, labels map[string]string, creationTime time.Time) *corev1.Pod {
+	return &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:              name,
+			Namespace:         namespace,
+			Labels:            labels,
+			CreationTimestamp: metav1.NewTime(creationTime),
+		},
 	}
 }
